Added constant.py, edited snake and qsnake
Updated snake,py in the following ways:
Added values gameWidth and gameHeight - these values are representative
of the game board. The window should be larger than the board in order
to display game information (the score). Default values are given.

Added text variable to game, this is a list of all text objects to be
drawn. Assumption is they derive from snake.DisplayText

Added new options to userInput - pressing p will cause the speed to slow
donw -- useful for debugging. Pressing o will reset the speed to 10.

drawBoard now draws the background of the game in black, and the
background of the information in purple.

Created class DisplayText that gives a basis for generating text for the
game. Text is white by default

Creatd class Score that inherits from DisplayText. Score is the numebr
of food eaten by the snake.

Snake starts off in the UP direction instead of NONE.

changeDirection has a new avoidFlip parameter. If we are changing the
direction through userInput, we can't flip around. However, if we are
assisting, the direction can be flipped.

snake now has instance variables hit_wall and hit_self instead of
viewing return from function.

die() doesn't change direction.

safeDirections will allow the assist to flip the direction if it is
safe.

assist() call smove after setting the new direction.

distanceToFood() added. Calculates distance from head of snake to food.